refactor(SkeletonItems): 미리 만들어둔 skeleton 리스트에 숨기고 보여줄 수 있도록 수정
