+++ src/components/SkeletonItems/skeleton-items.css
@@ -22,3 +22,18 @@
   animation: skeleton-animation 5s infinite ease-out;
   border-radius: 8px;
 }
+
+.hide-skeleton {
+  overflow: hidden;
+  visibility: hidden;
+
+  width: 0;
+  height: 0;
+}
+
+.skeleton-list {
+  display: grid;
+  grid-template-columns: repeat(4, 180px);
+  grid-column-gap: 160px;
+  grid-row-gap: 48px;
+}
